补全oss swift的demo
