+++ OSSSwiftDemo/OSSSwiftDemo/Classes/OSSSwiftGlobalDefines.swift
@@ -1,0 +1,23 @@
+//
+//  OSSSwiftGlobalDefines.swift
+//  OSSSwiftDemo
+//
+//  Created by 怀叙 on 2018/1/3.
+//  Copyright © 2018年 阿里云. All rights reserved.
+//
+
+import Foundation
+
+let OSS_ACCESSKEY_ID: String = "OSS_ACCESSKEY_ID"
+let OSS_SECRETKEY_ID: String = "OSS_SECRETKEY_ID"
+let OSS_BUCKET_PUBLIC: String = "huaixu-private"
+let OSS_BUCKET_PRIVATE: String = "huaixu-private"
+let OSS_ENDPOINT: String = "http://oss-cn-shenzhen.aliyuncs.com"
+let OSS_MULTIPART_UPLOADKEY: String = "multipart"
+let OSS_RESUMABLE_UPLOADKEY: String = "resumable"
+let OSS_CALLBACK_URL: String = "http://oss-demo.aliyuncs.com:23450"
+let OSS_CNAME_URL: String = "https://hs.chenhongyu.cn"
+let OSS_STSTOKEN_URL: String = "http://30.40.38.11:3015/sts/getsts"
+let OSS_IMAGE_KEY: String = "haskey.jpeg"
+
+
