--- OSSSwiftDemo/OSSSwiftDemo/MyObjcClass.h
@@ -1,15 +1,0 @@
-//
-//  MyObjcClass.h
-//  OSSSwiftDemo
-//
-//  Created by 怀叙 on 2018/1/2.
-//  Copyright © 2018年 阿里云. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-@interface MyObjcClass : NSObject
-
-- (void)sayHello:(NSString *)name;
-
-@end
