--- OSSSwiftDemo/OSSSwiftDemo/MyObjcClass.m
@@ -1,19 +1,0 @@
-//
-//  MyObjcClass.m
-//  OSSSwiftDemo
-//
-//  Created by 怀叙 on 2018/1/2.
-//  Copyright © 2018年 阿里云. All rights reserved.
-//
-
-#import "MyObjcClass.h"
-
-
-@implementation MyObjcClass
-
-- (void)sayHello:(NSString *)name
-{
-    
-}
-
-@end
