--- OSSSwiftDemo/OSSSwiftDemo/OSSNavigationController.swift
@@ -1,14 +1,0 @@
-//
-//  OSSNavigationController.swift
-//  OSSSwiftDemo
-//
-//  Created by 怀叙 on 2018/1/4.
-//  Copyright © 2018年 阿里云. All rights reserved.
-//
-
-import UIKit
-import Foundation
-
-class OSSNavigationController: UINavigationController {
-    
-}
