--- OSSSwiftDemo/OSSSwiftDemo/OSSSwiftGlobalDefines.swift
@@ -1,23 +1,0 @@
-//
-//  OSSSwiftGlobalDefines.swift
-//  OSSSwiftDemo
-//
-//  Created by 怀叙 on 2018/1/3.
-//  Copyright © 2018年 阿里云. All rights reserved.
-//
-
-import Foundation
-
-let OSS_ACCESSKEY_ID: String = "OSS_ACCESSKEY_ID"
-let OSS_SECRETKEY_ID: String = "OSS_SECRETKEY_ID"
-let OSS_BUCKET_PUBLIC: String = "OSS_BUCKET_PUBLIC"
-let OSS_BUCKET_PRIVATE: String = "OSS_BUCKET_PRIVATE"
-let OSS_ENDPOINT: String = "http://region.aliyuncs.com"
-let OSS_MULTIPART_UPLOADKEY: String = "OSS_MULTIPART_UPLOADKEY"
-let OSS_RESUMABLE_UPLOADKEY: String = "OSS_RESUMABLE_UPLOADKEY"
-let OSS_CALLBACK_URL: String = "OSS_CALLBACK_URL"
-let OSS_CNAME_URL: String = "OSS_CNAME_URL"
-let OSS_STSTOKEN_URL: String = "OSS_STSTOKEN_URL"
-let OSS_IMAGE_KEY: String = "OSS_IMAGE_KEY"
-
-
