+++ OSSSwiftDemo/OSSSwiftDemo/Resources/test.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string-array name="languages">
+        <item>C语言</item>
+        <item>Java </item>
+        <item>C#</item>
+        <item>PHP</item>
+        <item>HTML</item>
+    </string-array>
+    
+    <integer-array name="reminder_methods_values" translatable="false">
+        <item>1</item>
+        <item>2</item>
+        <item>3</item>
+    </integer-array>
+    
+    <note>
+        <to>George</to>
+        <from>John</from>
+        <heading>Reminder</heading>
+        <body>Don't forget the meeting!</body>
+    </note>
+    
+    <title>hello world</title>
+    
+    <script>
+        <![CDATA[
+            function matchwo(a,b)
+            {
+            if (a < b && a < 0) then
+            {
+            return 1
+            }
+            else
+            {
+            return 0
+            }
+            }
+        ]]>
+    </script>
+</resources>
+
